Move createImageTexture to media-utils
